move demos to rebranding scope
