+++ docs/src/components/home/Hero.tsx
@@ -6,19 +6,19 @@ import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import Stack from '@material-ui/core/Stack';
 import GradientText from 'docs/src/components/typography/GradientText';
-import PlayerCard from 'docs/src/pages/components/cards/PlayerCard';
-import TaskCard from 'docs/src/pages/components/cards/TaskCard';
-import NotificationCard from 'docs/src/pages/components/cards/NotificationCard';
-import ThemeChip from 'docs/src/pages/components/chips/ThemeChip';
-import ThemeDatePicker from 'docs/src/pages/components/date-picker/ThemeDatePicker';
-import ThemeSlider from 'docs/src/pages/components/slider/ThemeSlider';
-import FolderTable from 'docs/src/pages/components/tables/FolderTable';
-import ThemeTabs from 'docs/src/pages/components/tabs/ThemeTabs';
-import ThemeTimeline from 'docs/src/pages/components/timeline/ThemeTimeline';
-import ViewToggleButton from 'docs/src/pages/components/toggle-button/ViewToggleButton';
-import ThemeToggleButton from 'docs/src/pages/components/toggle-button/ThemeToggleButton';
-import ThemeSwitch from 'docs/src/pages/components/switches/ThemeSwitch';
-import ThemeAccordion from 'docs/src/pages/components/accordion/ThemeAccordion';
+import PlayerCard from 'docs/src/components/showcase/PlayerCard';
+import TaskCard from 'docs/src/components/showcase/TaskCard';
+import NotificationCard from 'docs/src/components/showcase/NotificationCard';
+import ThemeChip from 'docs/src/components/showcase/ThemeChip';
+import ThemeDatePicker from 'docs/src/components/showcase/ThemeDatePicker';
+import ThemeSlider from 'docs/src/components/showcase/ThemeSlider';
+import FolderTable from 'docs/src/components/showcase/FolderTable';
+import ThemeTabs from 'docs/src/components/showcase/ThemeTabs';
+import ThemeTimeline from 'docs/src/components/showcase/ThemeTimeline';
+import ViewToggleButton from 'docs/src/components/showcase/ViewToggleButton';
+import ThemeToggleButton from 'docs/src/components/showcase/ThemeToggleButton';
+import ThemeSwitch from 'docs/src/components/showcase/ThemeSwitch';
+import ThemeAccordion from 'docs/src/components/showcase/ThemeAccordion';
 import GetStartedButtons from 'docs/src/components/home/GetStartedButtons';
 
 export default function Hero() {
