+++ docs/src/components/showcase/NotificationCard.tsx
