+++ docs/src/components/showcase/ThemeAccordion.tsx
