+++ docs/src/components/showcase/ThemeSlider.tsx
