+++ docs/src/components/showcase/ThemeSwitch.tsx
