+++ docs/src/components/showcase/ViewToggleButton.tsx
